backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
tensorflow>=2.16.0; sys_platform != "darwin" or python_version >= "3.10"
# torch>=2.2.0  # uncomment if you prefer PyTorch
python-dotenv>=1.0.0
